Revert "refactor(localize): move `source_file_utils.ts` up (#36834)"
This reverts commit d669429bd26dbd753584c331a6d7aa44a4d1f3ee.

The Angular CLI relies upon deep imports into `@angular/localize`. In this
case it relies upon the `source_file_utils.ts` file being in its previous
position.

In `master` (i.e. v10) the CLI will update its import to cope but we need
to revert for 9.1.x.
