+++ packages/localize/src/tools/src/translate/source_files/es2015_translate_plugin.ts
@@ -11,7 +11,7 @@ import {TaggedTemplateExpression} from '@babel/types';
 
 import {Diagnostics} from '../../diagnostics';
 
-import {buildCodeFrameError, buildLocalizeReplacement, isBabelParseError, isLocalize, translate, TranslatePluginOptions, unwrapMessagePartsFromTemplateLiteral} from '../../source_file_utils';
+import {buildCodeFrameError, buildLocalizeReplacement, isBabelParseError, isLocalize, translate, TranslatePluginOptions, unwrapMessagePartsFromTemplateLiteral} from './source_file_utils';
 
 export function makeEs2015TranslatePlugin(
     diagnostics: Diagnostics, translations: Record<string, ɵParsedTranslation>,
