+++ packages/localize/src/tools/src/translate/source_files/source_file_utils.ts
@@ -8,7 +8,8 @@
 import {ɵisMissingTranslationError, ɵmakeTemplateObject, ɵParsedTranslation, ɵSourceLocation, ɵtranslate} from '@angular/localize';
 import {NodePath} from '@babel/traverse';
 import * as t from '@babel/types';
-import {Diagnostics} from './diagnostics';
+
+import {Diagnostics} from '../../diagnostics';
 
 /**
  * Is the given `expression` the global `$localize` identifier?
